Màj de la logique des méthodes "profileSettings" et "bannerSettings" : elles peuvent maintenant modifier une photo même si l'utilisateur n'en a aucune

+++ src/Controller/UserController.php
@@ -2,10 +2,12 @@
 
 namespace App\Controller;
 
+use App\Entity\Post;
 use App\Entity\User;
 use App\Form\UserSettingsType;
 use App\Repository\PostRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bridge\Doctrine\Attribute\MapEntity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
@@ -44,6 +46,15 @@ final class UserController extends AbstractController
         ]);
     }
 
+    #[Route('/user/{id}', name: 'show_profile')]
+    public function showPostProfile(Post $post)
+    {
+        $userPosts = $post->getMyUser();
+
+        return $this->render('post/show/showProfile.html.twig', [
+            'userPosts' => $userPosts
+        ]);
+    }
 
     /*
     La ligne "#[CurrentUser] User $user" va récupérer toutes les informations de l'utilisateur actuellement connecté :
@@ -52,8 +63,9 @@ final class UserController extends AbstractController
     - #[CurrentUser] : est un attribut PHP qui va me permettre de récupérer toutes les informations de l'utilisateur connecté
     */
     #[Route('/settings', name: 'settings')]
-    public function settings(#[CurrentUser] User $user): Response {
-        
+    public function settings(#[CurrentUser] User $user): Response
+    {
+
         $editUserForm = $this->createForm(UserSettingsType::class, $user);
 
         return $this->render('user/userSettings/settings.html.twig', [
@@ -77,34 +89,52 @@ final class UserController extends AbstractController
 
         // Si le formulaire est envoyé (isSubmitted) et que ses données sont valides (isValid())
         if ($editUserForm->isSubmitted() && $editUserForm->isValid()) {
-            
+
             // Récupère la valeur de l'input "profile_picture_name" et la stocke dans la variable $newProfilePicture
             $newProfilePicture = $editUserForm->get('profile_picture_name')->getData();
 
             if ($newProfilePicture) {
-                unlink($oldProfilePicture);
 
-                // Je récupère le nom original de l'image
-                $originalImageName = pathinfo($newProfilePicture->getClientOriginalName(), PATHINFO_FILENAME);
+                if (is_dir($oldProfilePicture)) {
 
-                // Puis j'utilise un slug pour créer un nom sécurisé
-                $safeFilename = $slugger->slug($originalImageName);
+                    $originalFilename = pathinfo($newProfilePicture->getClientOriginalName(), PATHINFO_FILENAME);
+                    $safeFilename = $slugger->slug($originalFilename);
+                    $newFilename = $safeFilename . '-' . uniqid() . '.' . $newProfilePicture->guessExtension();
 
-                /* Enfin, je crée le nom définitif de l'image en utilisant la valeur de la variable $safeFileName, j'inclus
-                un identifiant unique grâce à la fonction "uniqid()", puis je précise l'extension de l'image grâce à la fonction
-                "guessExtension()" appliquée sur la variable $image.
-                */
-                $newFilename = $safeFilename . '-' . uniqid() . '.' . $newProfilePicture->guessExtension();
+                    try {
+                        $newProfilePicture->move($profilePicturesDirectory, $newFilename);
+                    } catch (FileException $e) {
+                        $message = $e;
+                    }
 
-                try {
-                    // Envoie l'image dans le dossier adéquat
-                    $newProfilePicture->move($profilePicturesDirectory, $newFilename);
-                } catch (FileException $e) {
-                    $message = $e;
+                    // Met à jour le nom du fichier dans la BDD
+                    $user->setProfilePictureName($newFilename);
+
+                    // Sinon, si je récupère un fichier :
+                } else {
+                    unlink($oldProfilePicture);
+
+                    // Je récupère le nom original de l'image
+                    $originalImageName = pathinfo($newProfilePicture->getClientOriginalName(), PATHINFO_FILENAME);
+
+                    // Puis j'utilise un slug pour créer un nom sécurisé
+                    $safeFilename = $slugger->slug($originalImageName);
+
+                    /* Enfin, je crée le nom définitif de l'image en utilisant la valeur de la variable $safeFileName, j'inclus
+                    un identifiant unique grâce à la fonction "uniqid()", puis je précise l'extension de l'image grâce à la
+                    fonction "guessExtension()" appliquée sur la variable $image. */
+                    $newFilename = $safeFilename . '-' . uniqid() . '.' . $newProfilePicture->guessExtension();
+
+                    try {
+                        // Envoie l'image dans le dossier adéquat
+                        $newProfilePicture->move($profilePicturesDirectory, $newFilename);
+                    } catch (FileException $e) {
+                        $message = $e;
+                    }
+
+                    // Stocke le nom de l'image dans la BDD
+                    $user->setProfilePictureName($newFilename);
                 }
-
-                // Stocke le nom de l'image dans la BDD
-                $user->setProfilePictureName($newFilename);
             }
 
             $em->persist($user); // Prépare la requête
@@ -112,15 +142,16 @@ final class UserController extends AbstractController
 
             return $this->redirectToRoute('user_index');
         }
-        
+
         return $this->render('user/userSettings/pictures/profilePictureSettings.html.twig', [
             'editUserForm' => $editUserForm
         ]);
     }
 
     #[Route('/settings/banner', name: 'banner_settings')]
-    public function bannerSettings(#[CurrentUser] User $user, Request $request, EntityManagerInterface $em, SluggerInterface $slugger, #[Autowire('%kernel.project_dir%/public/uploads/user_images/banners')] string $bannersDirectory): Response {
-        
+    public function bannerSettings(#[CurrentUser] User $user, Request $request, EntityManagerInterface $em, SluggerInterface $slugger, #[Autowire('%kernel.project_dir%/public/uploads/user_images/banners')] string $bannersDirectory): Response
+    {
+
         /* J'ai ajouté un nouveau paramètre dans le fichier services.yaml nommé "user_images_directory", qui va contenir toutes
         les images associées à l'utilisateur */
         $oldBanner = $this->getParameter("user_images_directory") . '/banners/' . $user->getBannerName(); // Récupère la photo de profil actuelle
@@ -134,34 +165,51 @@ final class UserController extends AbstractController
 
         // Si le formulaire est envoyé (isSubmitted) et que ses données sont valides (isValid())
         if ($editUserForm->isSubmitted() && $editUserForm->isValid()) {
-            
+
             // Récupère la valeur de l'input "banner_name" et la stocke dans la variable $newBanner
             $newBanner = $editUserForm->get('banner_name')->getData();
 
             if ($newBanner) {
-                unlink($oldBanner);
 
-                // Je récupère le nom original de l'image
-                $originalImageName = pathinfo($newBanner->getClientOriginalName(), PATHINFO_FILENAME);
+                if (is_dir($oldBanner)) {
+                    
+                    $originalFilename = pathinfo($newBanner->getClientOriginalName(), PATHINFO_FILENAME);
+                    $safeFilename = $slugger->slug($originalFilename);
+                    $newFilename = $safeFilename . '-' . uniqid() . '.' . $newBanner->guessExtension();
 
-                // Puis j'utilise un slug pour créer un nom sécurisé
-                $safeFilename = $slugger->slug($originalImageName);
+                    try {
+                        $newBanner->move($bannersDirectory, $newFilename);
+                    } catch (FileException $e) {
+                        $message = $e;
+                    }
 
-                /* Enfin, je crée le nom définitif de l'image en utilisant la valeur de la variable $safeFileName, j'inclus
-                un identifiant unique grâce à la fonction "uniqid()", puis je précise l'extension de l'image grâce à la fonction
-                "guessExtension()" appliquée sur la variable $image.
-                */
-                $newFilename = $safeFilename . '-' . uniqid() . '.' . $newBanner->guessExtension();
+                    // Met à jour le nom du fichier dans la BDD
+                    $user->setBannerName($newFilename);
 
-                try {
-                    // Envoie l'image dans le dossier adéquat
-                    $newBanner->move($bannersDirectory, $newFilename);
-                } catch (FileException $e) {
-                    $message = $e;
+                } else {
+                    unlink($oldBanner);
+
+                    // Je récupère le nom original de l'image
+                    $originalImageName = pathinfo($newBanner->getClientOriginalName(), PATHINFO_FILENAME);
+
+                    // Puis j'utilise un slug pour créer un nom sécurisé
+                    $safeFilename = $slugger->slug($originalImageName);
+
+                    /* Enfin, je crée le nom définitif de l'image en utilisant la valeur de la variable $safeFileName, j'inclus
+                    un identifiant unique grâce à la fonction "uniqid()", puis je précise l'extension de l'image grâce à la
+                    fonction "guessExtension()" appliquée sur la variable $image. */
+                    $newFilename = $safeFilename . '-' . uniqid() . '.' . $newBanner->guessExtension();
+
+                    try {
+                        // Envoie l'image dans le dossier adéquat
+                        $newBanner->move($bannersDirectory, $newFilename);
+                    } catch (FileException $e) {
+                        $message = $e;
+                    }
+
+                    // Stocke le nom de l'image dans la BDD
+                    $user->setBannerName($newFilename);
                 }
-
-                // Stocke le nom de l'image dans la BDD
-                $user->setBannerName($newFilename);
             }
 
             $em->persist($user); // Prépare la requête
